requirements.txt
crewai
langchain
langchain_groq 
python-dotenv
PyGithub
google-generativeai
requests
crewai_tools
embedchain
sentence-transformers